refactor: 패키지 이름 및 각 클래스의 위치 변경
