+++ src/main/java/racing/domain/Car.java
@@ -1,4 +1,4 @@
-package racing.car;
+package racing.domain;
 
 public class Car implements ReadonlyCar {
