+++ src/main/java/racing/domain/Movement.java
@@ -1,4 +1,4 @@
-package racing.car;
+package racing.domain;
 
 public interface Movement {
