+++ src/test/java/racing/RacingTest.java
@@ -2,9 +2,10 @@ package racing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racing.car.Car;
-import racing.car.ReadOnlyCarProxy;
-import racing.car.ReadonlyCar;
+import racing.domain.Car;
+import racing.domain.Racing;
+import racing.domain.ReadOnlyCarProxy;
+import racing.domain.ReadonlyCar;
 import racing.dto.RacingReport;
 
 import java.util.Arrays;
