+++ src/test/java/racing/domain/CarTest.java
@@ -1,4 +1,4 @@
-package racing.car;
+package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
