+++ src/test/java/racing/domain/RandomMovementTest.java
@@ -1,4 +1,4 @@
-package racing.car;
+package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
